feat(otel): Add support for Context.makeCurrent()
